chore(upgrade): Fixed vantage tests, refactored server relations out of common to avoid import crossover, fixed polyfills
