+++ src/server/registry/entityRegistry.spec.ts
@@ -1,8 +1,9 @@
-import { EntityRegistry, EntityType, registry } from './entityRegistry';
-import { Model, Controller, Seeder, Migration, Store, Service } from './decorators';
-import { AbstractModel } from '../models/model';
+import { EntityRegistry, EntityType, registry } from '../../common/registry/entityRegistry';
+import { Model, Store, Service } from '../../common/registry/decorators';
+import { AbstractModel } from '../../common/models/model';
+import { Controller, Seeder, Migration } from './decorators';
 
-describe('Entity registry', () => {
+fdescribe('Entity registry', () => {
 
   class Foo {
   }
@@ -193,14 +194,14 @@ describe('Entity registry', () => {
 
       let type = (decorator.toLowerCase() as EntityType);
 
-      it(`decorates class with @${decorator} to register class as tye '${type}'`, () => {
+      it(`decorates class with @${decorator} to register class as type '${type}'`, () => {
 
         let className  = 'Foo' + decorator;
         let foundClass = registry.getAllOfType(type)
           .get(className);
 
         expect(foundClass instanceof Function)
-          .toBe(true);
+          .toBe(true, typeof foundClass);
         expect(foundClass.name)
           .toEqual(className);
